refactor: extract reducer content operations
